requirements.txt
python-dotenv
pyautogui
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
